+++ app/frontend/src/types/chatting.ts
@@ -1,0 +1,11 @@
+export type Talk = {
+  id: string;
+  user: User;
+  datetime: Date;
+  content: string;
+};
+
+export type User = {
+  username: string;
+  profileSrc: string;
+};
