+++ app/frontend/src/types/index.ts
@@ -1,12 +1,1 @@
-// export * from "";
-export type User = {
-  username: string;
-  profileSrc: string;
-};
-
-export type Talk = {
-  id: string;
-  user: User;
-  datetime: Date;
-  content: string;
-};
+export * from './chatting';
